# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShilpSetu API",
    "description": "One photo. One spoken sentence. A fair-priced, trust-verified listing — no typing, no middleman, no English required. · SIH 2026 · PS ID SIH26090 · Team HACKER LOBBY",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/cart": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Get Cart",
        "operationId": "get_cart_api_v1_cart_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Add To Cart",
        "operationId": "add_to_cart_api_v1_cart_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartLine"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cart/{product_id}": {
      "delete": {
        "tags": [
          "commerce"
        ],
        "summary": "Remove From Cart",
        "operationId": "remove_from_cart_api_v1_cart__product_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/checkout": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Checkout",
        "operationId": "checkout_api_v1_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/orders": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "My Orders",
        "operationId": "my_orders_api_v1_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_v1_orders__order_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artisan/orders": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Artisan Orders",
        "operationId": "artisan_orders_api_v1_artisan_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-artisan-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Artisan-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/status": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Update Status",
        "description": "Artisans accept / decline (one tap or IVR key), pack and ship; buyers cancel; courier/admin mark delivered.",
        "operationId": "update_status_api_v1_orders__order_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/confirm": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Confirm Payment",
        "operationId": "confirm_payment_api_v1_payments_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/payments/mock/{provider_order_id}/pay": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Mock Pay",
        "description": "Dev-only UPI sandbox: simulates the buyer completing payment and returns what the SDK would return.",
        "operationId": "mock_pay_api_v1_payments_mock__provider_order_id__pay_post",
        "parameters": [
          {
            "name": "provider_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/webhook": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Payment Webhook",
        "operationId": "payment_webhook_api_v1_payments_webhook_post",
        "parameters": [
          {
            "name": "x-razorpay-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Razorpay-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Add Review",
        "operationId": "add_review_api_v1_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reviews/{product_id}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "List Reviews",
        "operationId": "list_reviews_api_v1_reviews__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Track Events",
        "description": "Buyer views / add-to-cart / orders / returns — these retrain the price and ranking models.",
        "operationId": "track_events_api_v1_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "line1": {
            "type": "string",
            "title": "Line1"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "pincode": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Pincode"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone",
          "line1",
          "city",
          "state",
          "pincode"
        ],
        "title": "Address"
      },
...
      "CartLine": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "product_id"
        ],
        "title": "CartLine"
      },
      "CheckoutIn": {
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "upi",
              "cod"
            ],
            "title": "Payment Method",
            "default": "upi"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "CheckoutIn"
      },
...
      "EventIn": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "view",
              "add_to_cart",
              "order",
              "return"
            ],
            "title": "Type"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "type",
          "product_id"
        ],
        "title": "EventIn"
      },
      "EventsIn": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventIn"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "EventsIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderStatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "declined",
              "packed",
              "shipped",
              "delivered",
              "cancelled",
              "returned"
            ],
            "title": "Status"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OrderStatusIn"
      },
...
      "PaymentConfirm": {
        "properties": {
          "provider_order_id": {
            "type": "string",
            "title": "Provider Order Id"
          },
          "payment_id": {
            "type": "string",
            "title": "Payment Id"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "provider_order_id",
          "payment_id",
          "signature"
        ],
        "title": "PaymentConfirm"
      },
...
      "ReviewIn": {
        "properties": {
          "order_item_id": {
            "type": "string",
            "title": "Order Item Id"
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "order_item_id",
          "rating"
        ],
        "title": "ReviewIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}